refactor: decompose oversized workflow loader.go (#661)
Split internal/workflow/loader.go (852 lines, over the 800-line production
file budget) into cohesive same-package files. Pure behavior-preserving move
— no signature, logic, or package-boundary changes:

  - errors.go       error categories, sentinels, Error/TemplateParseError
  - frontmatter.go  splitFrontMatter, hasNestedKey, hookFieldPresence
  - migrate.go      tracker/polling deprecation migrations + continuation-turn default
  - reject.go       removed-field rejection helpers + unknown-key logging
  - expand.go       config expansion / default-value assignment
  - env.go          ${VAR}/$VAR resolution + MissingEnvValueError
  - normalize.go    per-state concurrency cap normalization

Validation lookup tables (supported* maps) and isLinearGraphQLMutationName
move into validate.go next to their only consumer (validateConfig). loader.go
keeps load orchestration only (Load/LoadOptional/validateFrontMatterRoot).

Removes the internal/workflow/loader.go entry (852) from the file-size budget
baseline; loader.go is now 123 lines. Adds migrate_test.go characterization
tests (mutation-verified) pinning the gitea-skip, default-branch, and nil-safe
continuation-turn defaulting branches the existing suite covered only
indirectly.

Closes #661 for this file.
