+++ internal/workflow/loader.go
@@ -3,12 +3,8 @@ package workflow
 import (
 	"errors"
 	"fmt"
-	"log"
 	"os"
-	"path/filepath"
-	"sort"
 	"strings"
-	"time"
 
 	"gopkg.in/yaml.v3"
 )
@@ -18,117 +14,6 @@ type Workflow struct {
 	Config         Config
 	PromptTemplate string
 	Source         Source
-}
-
-type Category string
-
-const (
-	CategoryMissingWorkflowFile       Category = "missing_workflow_file"
-	CategoryWorkflowParseError        Category = "workflow_parse_error"
-	CategoryWorkflowFrontMatterNotMap Category = "workflow_front_matter_not_a_map"
-	CategoryTemplateParseError        Category = "template_parse_error"
-	CategoryTemplateRenderError       Category = "template_render_error"
-)
-
-var (
-	ErrMissingWorkflowFile       = &Error{Category: CategoryMissingWorkflowFile}
-	ErrWorkflowParse             = &Error{Category: CategoryWorkflowParseError}
-	ErrWorkflowFrontMatterNotMap = &Error{Category: CategoryWorkflowFrontMatterNotMap}
-	ErrTemplateParse             = &TemplateParseError{}
-	ErrTemplateRender            = &TemplateRenderError{}
-)
-
-type Error struct {
-	Category Category
-	Path     string
-	Message  string
-	Err      error
-}
-
-func NewError(category Category, message string, err error) *Error {
-	return &Error{Category: category, Message: message, Err: err}
-}
-
-func (e *Error) Error() string {
-	if e == nil {
-		return ""
-	}
-	msg := e.Message
-	if msg == "" {
-		msg = string(e.Category)
-	}
-	if e.Path != "" {
-		msg = e.Path + ": " + msg
-	}
-	if e.Err != nil {
-		msg += ": " + e.Err.Error()
-	}
-	return msg
-}
-
-func (e *Error) Unwrap() error {
-	if e == nil {
-		return nil
-	}
-	return e.Err
-}
-
-func (e *Error) Is(target error) bool {
-	return categoryMatches(e.category(), target)
-}
-
-func (e *Error) category() Category {
-	if e == nil {
-		return ""
-	}
-	return e.Category
-}
-
-type TemplateParseError struct {
-	Err error
-}
-
-func (e *TemplateParseError) Error() string {
-	if e == nil || e.Err == nil {
-		return string(CategoryTemplateParseError)
-	}
-	return fmt.Sprintf("%s: %v", CategoryTemplateParseError, e.Err)
-}
-
-func (e *TemplateParseError) Unwrap() error {
-	if e == nil {
-		return nil
-	}
-	return e.Err
-}
-
-func (e *TemplateParseError) Is(target error) bool {
-	return categoryMatches(e.category(), target)
-}
-
-func (e *TemplateParseError) category() Category {
-	return CategoryTemplateParseError
-}
-
-type categorized interface {
-	category() Category
-}
-
-func ErrorCategory(err error) (Category, bool) {
-	var categorizedErr categorized
-	if errors.As(err, &categorizedErr) {
-		category := categorizedErr.category()
-		return category, category != ""
-	}
-	return "", false
-}
-
-func categoryMatches(category Category, target error) bool {
-	if category == "" || target == nil {
-		return false
-	}
-	var targetCategory categorized
-	return errors.As(target, &targetCategory) && targetCategory.category() == category
 }
 
 func Load(path string) (*Workflow, error) { //nolint:gocognit // baseline (#521)
@@ -215,312 +100,6 @@ func validateFrontMatterRoot(path string, frontBytes []byte) error {
 	return nil
 }
 
-// migrateTrackerEndpoint reconciles the SPEC §5.3.1 `tracker.endpoint`
-// field with the pre-#242 `tracker.base_url` alias: prefer endpoint when
-// set, fall back to base_url with a deprecation log, and finally clear
-// BaseURL on the resolved config so downstream code reads only Endpoint.
-func migrateTrackerEndpoint(frontBytes []byte, cfg *Config) {
-	endpointPresent := hasNestedKey(frontBytes, "tracker", "endpoint")
-	legacyPresent := hasNestedKey(frontBytes, "tracker", "base_url")
-	switch {
-	case endpointPresent:
-		if legacyPresent {
-			log.Printf("workflow: tracker.base_url is deprecated and ignored because tracker.endpoint is set")
-		}
-	case legacyPresent:
-		log.Printf("workflow: tracker.base_url is deprecated; use tracker.endpoint (SPEC §5.3.1)")
-		cfg.Tracker.Endpoint = cfg.Tracker.BaseURL
-	}
-	cfg.Tracker.BaseURL = ""
-}
-
-func migrateGiteaTrackerProjectSlug(frontBytes []byte, cfg *Config) {
-	if !strings.EqualFold(cfg.Tracker.Kind, "gitea") {
-		return
-	}
-	legacyPresent := hasNestedKey(frontBytes, "tracker", "project_slug")
-	if !legacyPresent {
-		return
-	}
-	switch {
-	case strings.TrimSpace(cfg.Tracker.Endpoint) != "":
-		log.Printf("workflow: tracker.project_slug is ignored for Gitea because tracker.endpoint is already set")
-	case strings.TrimSpace(cfg.Tracker.ProjectSlug) != "":
-		log.Printf("workflow: tracker.project_slug is deprecated as a Gitea base URL; use tracker.endpoint")
-		cfg.Tracker.Endpoint = cfg.Tracker.ProjectSlug
-	}
-	cfg.Tracker.ProjectSlug = ""
-}
-
-func migratePollingInterval(frontBytes []byte, cfg *Config) {
-	pollingPresent := hasNestedKey(frontBytes, "polling", "interval_ms")
-	legacyPresent := hasNestedKey(frontBytes, "tracker", "poll_interval_ms")
-	switch {
-	case pollingPresent:
-		if legacyPresent {
-			log.Printf("workflow: tracker.poll_interval_ms is deprecated and ignored because polling.interval_ms is set")
-		}
-		cfg.Tracker.PollIntervalMs = cfg.Polling.IntervalMs
-	case legacyPresent:
-		log.Printf("workflow: tracker.poll_interval_ms is deprecated; use polling.interval_ms")
-		cfg.Polling.IntervalMs = cfg.Tracker.PollIntervalMs
-	default:
-		cfg.Tracker.PollIntervalMs = cfg.Polling.IntervalMs
-	}
-}
-
-// supportedAgentDefaults mirrors the runner registry in
-// internal/runner.New. Keeping the two lists in sync at the schema layer
-// turns "unknown runner: X" — which today only surfaces after a task is
-// claimed and the workspace prepared — into a load-time configuration
-// error with the workflow file path attached.
-var supportedAgentDefaults = map[string]struct{}{
-	"mock":             {},
-	"codex-app-server": {},
-	"claude":           {},
-}
-
-var supportedSandboxBackends = map[string]struct{}{
-	"none":       {},
-	"bubblewrap": {},
-	"firejail":   {},
-}
-
-var supportedSandboxNetworks = map[string]struct{}{
-	"none":      {},
-	"allowlist": {},
-}
-
-// rejectRemovedFields surfaces a clear error for keys that were once part
-// of the schema but have been removed. The typed Unmarshal above silently
-// drops unknown fields, which would let workflow authors keep believing
-// the key still controls behavior. Targeted detection keeps existing
-// benign extras working while flagging known footguns.
-func hasNestedKey(front []byte, path ...string) bool {
-	var raw map[string]any
-	if err := yaml.Unmarshal(front, &raw); err != nil {
-		return false
-	}
-	var current any = raw
-	for _, key := range path {
-		m, ok := current.(map[string]any)
-		if !ok {
-			return false
-		}
-		current, ok = m[key]
-		if !ok {
-			return false
-		}
-	}
-	return true
-}
-
-func hookFieldPresence(front []byte, path ...string) HookFieldPresence {
-	return HookFieldPresence{
-		AfterCreate:    hasNestedKey(front, append(path, "after_create")...),
-		BeforeRun:      hasNestedKey(front, append(path, "before_run")...),
-		AfterRun:       hasNestedKey(front, append(path, "after_run")...),
-		BeforeRemove:   hasNestedKey(front, append(path, "before_remove")...),
-		TimeoutMs:      hasNestedKey(front, append(path, "timeout_ms")...),
-		EnvPassthrough: hasNestedKey(front, append(path, "env_passthrough")...),
-	}
-}
-
-func rejectRemovedFields(front []byte) error {
-	var raw map[string]any
-	if err := yaml.Unmarshal(front, &raw); err != nil {
-		return nil
-	}
-	if err := rejectRemovedCodexFields(raw); err != nil {
-		return err
-	}
-	if err := rejectRemovedVerifyFields(raw); err != nil {
-		return err
-	}
-	if err := rejectRemovedPolicyFields(raw); err != nil {
-		return err
-	}
-	return rejectRemovedMiscFields(raw)
-}
-
-// rejectRemovedMiscFields surfaces clear errors for removed keys that do not
-// have their own dedicated helper: the nested `claude.profile` / `agent.fallback`
-// keys removed earlier, and the top-level `pr:` / `safety:` blocks removed under
-// #578. Collecting them here keeps rejectRemovedFields a flat list of helper
-// calls (gocognit). Like the verify.*/policy.*/codex.* rejects, each fails loud
-// instead of being silently dropped. The `pr:`/`safety:` blocks configured
-// capabilities that no longer exist — the worker does not open/label PRs (PR
-// handoff is agent-side per SPEC §1, #76), and `safety:` was a descriptive
-// struct the worker never enforced — so their intent belongs in the WORKFLOW
-// prompt (SPEC §3.2), not front matter.
-func rejectRemovedMiscFields(raw map[string]any) error {
-	if claude, ok := raw["claude"].(map[string]any); ok {
-		if _, present := claude["profile"]; present {
-			return fmt.Errorf("claude.profile is not supported; the codex-only `profile` field was removed in issue #541 and Claude never had runner profiles. Remove it")
-		}
-	}
-	// pr:/safety: are top-level keys, so they must be checked regardless of
-	// whether an `agent:` block is present (rejectRemovedAgentFields below
-	// returns early when it is not).
-	if _, present := raw["pr"]; present {
-		return fmt.Errorf("pr.* is no longer supported (#578): the worker does not open, label, or assign reviewers to PRs — PR creation is the coding agent's responsibility (SPEC §1, #76). Express draft/labels/reviewers in the agent's WORKFLOW prompt or tool surface, and remove the `pr:` block")
-	}
-	if _, present := raw["safety"]; present {
-		return fmt.Errorf("safety.* is no longer supported (#578): it was a descriptive struct the worker never enforced, which falsely implied enforcement. Express the safety envelope in the WORKFLOW prompt/README, and use `sandbox:` for worker-enforced write/network restrictions. Remove the `safety:` block")
-	}
-	return rejectRemovedAgentFields(raw)
-}
-
-// rejectRemovedAgentFields surfaces clear errors for removed agent.* keys so a
-// stale workflow fails loud instead of being silently dropped: `agent.fallback`
-// (#40, never read) and the `agent.max_retry_attempts` / `agent.max_timeout_retries`
-// retry caps removed under #577 (the orchestrator no longer caps retries). Split
-// out of rejectRemovedFields to keep that function under the gocognit budget.
-func rejectRemovedAgentFields(raw map[string]any) error {
-	agent, ok := raw["agent"].(map[string]any)
-	if !ok {
-		return nil
-	}
-	if _, present := agent["fallback"]; present {
-		return fmt.Errorf("agent.fallback is no longer supported (issue #40); the worker never read this field. Remove it and set agent.default to a more reliable runner if you need a different default")
-	}
-	for _, key := range []string{"max_retry_attempts", "max_timeout_retries"} {
-		if _, present := agent[key]; present {
-			return fmt.Errorf("agent.%s is no longer supported (#577): the orchestrator no longer caps retries. SPEC §8.4 / §16.6 retry unboundedly with exponential backoff until the tracker takes the issue out of active work; to stop a persistently-failing issue, move it out of the active states. Remove agent.%s", key, key)
-		}
-	}
-	return nil
-}
-
-func defaultAgentMaxContinuationTurns(cfg *Config, hasFrontMatter bool, front []byte) {
-	if cfg == nil {
-		return
-	}
-	if hasFrontMatter && hasNestedKey(front, "agent", "max_continuation_turns") {
-		return
-	}
-	cfg.Agent.MaxContinuationTurns = cfg.Agent.MaxTurns
-}
-
-// rejectRemovedVerifyFields surfaces clear errors for the verify.* keys that
-// configured removed worker gates so a stale workflow fails loud instead of
-// silently dropping a control the operator believes is active:
-//
-//   - timeout / allow_failure / env_passthrough configured the removed worker
-//     verify gate (#557 / DEVIATIONS D33). Verification is the coding agent's
-//     job now — verify.commands are surfaced to the agent's prompt and the
-//     worker no longer runs them.
-//   - secret_scan configured the removed worker secret-scan gate (#561). It
-//     ran after the agent had already pushed, so it could only flag, never
-//     prevent, a leak, and it raced reconcile-cancel the same way the verify
-//     gate did (#557). PR CI / the host's secret scanning is the backstop now.
-//
-// verify.commands remains valid and is not rejected.
-func rejectRemovedVerifyFields(raw map[string]any) error {
-	verify, ok := raw["verify"].(map[string]any)
-	if !ok {
-		return nil
-	}
-	for _, key := range []string{"timeout", "allow_failure", "env_passthrough"} {
-		if _, present := verify[key]; present {
-			return fmt.Errorf("verify.%s is no longer supported (#557): the worker verify gate was removed, so verification is the agent's responsibility — verify.commands are surfaced to the prompt and the worker no longer runs them. Remove verify.%s and express the check in the WORKFLOW prompt instead", key, key)
-		}
-	}
-	if _, present := verify["secret_scan"]; present {
-		return fmt.Errorf("verify.secret_scan is no longer supported (#561): the worker secret-scan gate was removed — it ran after the agent had already pushed, so it could only flag, never prevent, a leak, and it raced reconcile-cancel. Remove verify.secret_scan and rely on PR CI / your host's secret scanning, or run the scan from the WORKFLOW prompt before the agent pushes")
-	}
-	return nil
-}
-
-// rejectRemovedPolicyFields surfaces clear errors for the policy path/diffstat
-// gate keys removed under #561, so a stale workflow fails loud instead of
-// silently dropping a scope/path control the operator believes is enforced. The
-// worker policy gate ran after the agent had already pushed (could only flag,
-// never prevent) and raced reconcile-cancel; SPEC §3.2 homes scope/validation
-// rules in the WORKFLOW.md prompt and upstream has no such config. policy.mode
-// remains valid (it selects draft_pr vs analysis_only).
-func rejectRemovedPolicyFields(raw map[string]any) error {
-	if policy, ok := raw["policy"].(map[string]any); ok {
-		for _, key := range []string{"allow_paths", "deny_paths", "max_changed_files", "max_changed_lines", "max_changed_loc"} {
-			if _, present := policy[key]; present {
-				return fmt.Errorf("policy.%s is no longer supported (#561): the worker path/diffstat gate was removed — it ran after the agent had already pushed, so it could only flag, never prevent. Express scope/path rules in the WORKFLOW prompt (SPEC §3.2) and use sandbox write restrictions for hard path prevention. Remove policy.%s; policy.mode is still honored", key, key)
-			}
-		}
-	}
-	if agent, ok := raw["agent"].(map[string]any); ok {
-		if _, present := agent["policy_violation_budget"]; present {
-			return fmt.Errorf("agent.policy_violation_budget is no longer supported (#561): the worker policy gate and its violation-retry loop were removed. Remove agent.policy_violation_budget")
-		}
-	}
-	return nil
-}
-
-// rejectRemovedCodexFields surfaces clear errors for codex.* keys that
-// configured the removed one-shot `codex exec` runner (issue #541): the
-// `codex.profile` field and a `codex.command` whose argv runs `codex exec`.
-// Both would otherwise be silently dropped or fall through to a `sh -c
-// "codex exec"` launch that the app-server runner cannot speak, failing the
-// first real run with an opaque protocol/timeout error.
-func rejectRemovedCodexFields(raw map[string]any) error {
-	codex, ok := raw["codex"].(map[string]any)
-	if !ok {
-		return nil
-	}
-	if _, present := codex["profile"]; present {
-		return fmt.Errorf("codex.profile is no longer supported (issue #541); the one-shot `codex exec` runner it configured was removed. The SPEC §10 runner is `codex app-server` (agent.default: codex-app-server); set its sandbox with codex.thread_sandbox / codex.turn_sandbox_policy instead")
-	}
-	if command, ok := codex["command"].(string); ok && commandRunsCodexExec(command) {
-		return fmt.Errorf("codex.command %q runs the removed one-shot `codex exec` runner (issue #541); the SPEC §10 runner is `codex app-server` — set codex.command to `codex app-server`", command)
-	}
-	return nil
-}
-
-// commandRunsCodexExec reports whether a codex.command launches the removed
-// one-shot `codex exec` runner (issue #541). It is a deliberately conservative
-// load-time guard: quote characters are stripped before tokenizing so quoted
-// spellings (e.g. `"codex" "exec"`) are caught too. The runner's
-// splitAppServerCommand is the authoritative launch-time parser — this only
-// exists to turn the otherwise opaque app-server JSON-RPC handshake failure into
-// a clear config error at load. Over-rejecting a pathological quoted command is
-// the safe direction (fail fast) versus silently launching the removed runner.
-func commandRunsCodexExec(command string) bool {
-	unquoted := strings.NewReplacer(`"`, "", `'`, "").Replace(command)
-	fields := strings.Fields(unquoted)
-	return len(fields) >= 2 && fields[0] == "codex" && fields[1] == "exec"
-}
-
-var knownTopLevelWorkflowKeys = map[string]struct{}{
-	"agent":     {},
-	"claude":    {},
-	"codex":     {},
-	"hooks":     {},
-	"policy":    {},
-	"polling":   {},
-	"repo":      {},
-	"sandbox":   {},
-	"server":    {},
-	"tracker":   {},
-	"verify":    {},
-	"workspace": {},
-}
-
-func logUnknownTopLevelKeys(front []byte) {
-	var raw map[string]any
-	if err := yaml.Unmarshal(front, &raw); err != nil {
-		return
-	}
-	unknown := make([]string, 0)
-	for key := range raw {
-		if _, ok := knownTopLevelWorkflowKeys[key]; !ok {
-			unknown = append(unknown, key)
-		}
-	}
-	sort.Strings(unknown)
-	for _, key := range unknown {
-		log.Printf("workflow: unknown top-level key %s ignored", key)
-	}
-}
-
 // LoadOptional loads a workflow from an explicit path, returning schema
 // defaults when the file does not exist. New worker code should use
 // Resolve(workdir), which handles repo-relative discovery and returns
@@ -541,312 +120,4 @@ func LoadOptional(path string) (*Workflow, error) {
 		return &Workflow{Path: path, Config: cfg, PromptTemplate: DefaultPrompt()}, nil
 	}
 	return nil, err
-}
-
-// splitFrontMatter peels the SPEC §5.2 YAML front-matter block off
-// the start of a workflow file. The opening fence is `---` followed
-// by a newline; the closing fence is a line that is **exactly**
-// `---` (with an optional CR before the LF) and nothing else. The
-// earlier substring-based scan would mis-match `---` lines that
-// appear inside YAML block scalars or quoted strings — see #231,
-// where `description: |` blocks legitimately contained a `---` line
-// and silently truncated the parsed Config.
-//
-// Returns (front, body). When no opening fence is present or no
-// closing fence can be found, returns ("", s) so the caller treats
-// the whole file as the prompt body.
-func splitFrontMatter(s string) (string, string) {
-	if !strings.HasPrefix(s, "---\n") && !strings.HasPrefix(s, "---\r\n") {
-		return "", s
-	}
-	rest := strings.TrimPrefix(strings.TrimPrefix(s, "---\r\n"), "---\n")
-	lines := strings.SplitAfter(rest, "\n")
-	var front strings.Builder
-	for i, line := range lines {
-		// Strip only the line-ending (CR/LF) for the fence comparison.
-		// Trailing spaces, indentation, or any other content disqualify
-		// the line from being the closing fence.
-		if strings.TrimRight(line, "\r\n") == "---" {
-			return front.String(), strings.Join(lines[i+1:], "")
-		}
-		front.WriteString(line)
-	}
-	return "", s
-}
-
-func expandConfig(cfg *Config) error {
-	return expandConfigForWorkflowPath("", cfg)
-}
-
-func expandConfigForWorkflowPath(workflowPath string, cfg *Config) error { //nolint:gocognit,funlen // baseline (#521)
-	var err error
-	if envName, ok := explicitEnvReferenceName(cfg.Tracker.APIKey); ok {
-		cfg.Tracker.apiKeyEnvVar = envName
-	} else {
-		cfg.Tracker.apiKeyEnvVar = ""
-	}
-	if cfg.Tracker.APIKey, err = resolveExplicitEnv("tracker.api_key", cfg.Tracker.APIKey); err != nil {
-		return err
-	}
-	if cfg.Tracker.Endpoint, err = resolveExplicitEnv("tracker.endpoint", cfg.Tracker.Endpoint); err != nil {
-		return err
-	}
-	if err := expandRepoConfig("repo.clone_url", &cfg.Repo); err != nil {
-		return err
-	}
-	if cfg.Workspace.Root, err = normalizeWorkflowPath(workflowPath, cfg.Workspace.Root); err != nil {
-		return err
-	}
-	if cfg.Codex.Command, err = resolveExplicitEnv("codex.command", cfg.Codex.Command); err != nil {
-		return err
-	}
-	if cfg.Claude.Command, err = resolveExplicitEnv("claude.command", cfg.Claude.Command); err != nil {
-		return err
-	}
-	for i := range cfg.Sandbox.CredentialFiles {
-		field := fmt.Sprintf("sandbox.credential_files[%d]", i)
-		resolved, err := resolveExplicitEnv(field, cfg.Sandbox.CredentialFiles[i])
-		if err != nil {
-			return err
-		}
-		cfg.Sandbox.CredentialFiles[i] = expandPath(resolved)
-	}
-	if cfg.Agent.Default == "" {
-		cfg.Agent.Default = "mock"
-	}
-	// agent.max_concurrent_agents: SPEC §6.4 default of 10 is supplied by
-	// DefaultConfig() and survives YAML overlay when the field is absent.
-	// An explicit `max_concurrent_agents: 0` (or any non-positive value)
-	// is rejected by validateConfig rather than silently coerced — Elixir
-	// `validate_number(:max_concurrent_agents, greater_than: 0)`
-	// (schema.ex:131,145) makes 0 a validation error, not a request for
-	// the default.
-	if cfg.Agent.Timeout <= 0 {
-		cfg.Agent.Timeout = 30 * time.Minute
-	}
-	if cfg.Polling.IntervalMs <= 0 {
-		cfg.Polling.IntervalMs = cfg.Tracker.PollIntervalMs
-	}
-	if cfg.Polling.IntervalMs <= 0 {
-		cfg.Polling.IntervalMs = 30000
-	}
-	cfg.Tracker.PollIntervalMs = cfg.Polling.IntervalMs
-	// Tracker.Statuses defaults are applied per-field so a YAML override of
-	// a single name (e.g. `statuses.in_progress: "Doing"`) does not require
-	// the operator to also restate the unchanged ones. The defaults match
-	// the Linear template the personal profile ships with.
-	if cfg.Tracker.Statuses.InProgress == "" {
-		cfg.Tracker.Statuses.InProgress = "In Progress"
-	}
-	if cfg.Tracker.Statuses.HumanReview == "" {
-		cfg.Tracker.Statuses.HumanReview = "Human Review"
-	}
-	if cfg.Tracker.Statuses.Rework == "" {
-		cfg.Tracker.Statuses.Rework = "Rework"
-	}
-	if cfg.Sandbox.Backend == "" {
-		cfg.Sandbox.Backend = "none"
-	}
-	if cfg.Sandbox.NetworkMode == "" {
-		cfg.Sandbox.NetworkMode = "none"
-	}
-	if cfg.Codex.ThreadSandbox == "" {
-		cfg.Codex.ThreadSandbox = "workspace-write"
-	}
-	// Derive the per-turn sandbox policy from thread_sandbox unless the
-	// operator set codex.turn_sandbox_policy explicitly. ThreadSandbox is
-	// defaulted just above, so this sees the resolved thread sandbox. Deriving
-	// here (rather than pinning workspace-write in DefaultConfig) keeps
-	// thread_sandbox as the single knob governing effective turn permissions
-	// (#472; DEVIATIONS D32).
-	if shouldDeriveTurnSandboxPolicy(cfg.Codex) {
-		cfg.Codex.TurnSandboxPolicy = defaultTurnSandboxPolicyForThread(cfg.Codex.ThreadSandbox)
-	}
-	if cfg.Codex.ApprovalPolicy == nil {
-		// Default mirrors codex's `granular` policy with every flag set to
-		// false, i.e. "auto-reject every approval / elicitation prompt".
-		// Codex renamed the variant from `reject` → `granular` and flipped
-		// the field polarity (true = allow) in
-		// codex commit b7dba72db (#14516); aiops-platform had been sending
-		// the obsolete `reject:` payload, which made every thread/start
-		// return `-32600 unknown variant ` (#329).
-		cfg.Codex.ApprovalPolicy = map[string]any{"granular": map[string]any{
-			"sandbox_approval":    false,
-			"rules":               false,
-			"skill_approval":      false,
-			"request_permissions": false,
-			"mcp_elicitations":    false,
-		}}
-	}
-	return nil
-}
-
-func expandRepoConfig(field string, repo *RepoConfig) error {
-	var err error
-	if repo.CloneURL, err = resolveExplicitEnv(field, repo.CloneURL); err != nil {
-		return err
-	}
-	if repo.DefaultBranch == "" {
-		repo.DefaultBranch = "main"
-	}
-	return nil
-}
-
-func normalizeWorkflowPath(workflowPath, p string) (string, error) {
-	resolved, err := resolveExplicitEnv("workspace.root", p)
-	if err != nil {
-		return "", err
-	}
-	expanded := expandPath(resolved)
-	if expanded == "" || filepath.IsAbs(expanded) || workflowPath == "" {
-		return expanded, nil
-	}
-	log.Printf("workflow: relative workspace.root %s resolved relative to workflow file %s", expanded, workflowPath)
-	return filepath.Join(filepath.Dir(workflowPath), expanded), nil
-}
-
-func expandPath(p string) string {
-	if p == "" {
-		return p
-	}
-	if strings.HasPrefix(p, "~/") {
-		home, err := os.UserHomeDir()
-		if err == nil {
-			return filepath.Join(home, strings.TrimPrefix(p, "~/"))
-		}
-	}
-	return p
-}
-
-type MissingEnvValueError struct {
-	Field  string
-	EnvVar string
-}
-
-func (e *MissingEnvValueError) Error() string {
-	category := "workflow_config_missing_value"
-	if e.Field == "tracker.api_key" {
-		category = "missing_tracker_api_key"
-	}
-	return fmt.Sprintf("%s: %s references $%s but the environment variable is unset or empty", category, e.Field, e.EnvVar)
-}
-
-func resolveExplicitEnv(field, value string) (string, error) {
-	envName, ok := explicitEnvReferenceName(value)
-	if !ok {
-		return value, nil
-	}
-	resolved, ok := os.LookupEnv(envName)
-	if !ok || resolved == "" {
-		return "", &MissingEnvValueError{Field: field, EnvVar: envName}
-	}
-	return resolved, nil
-}
-
-func explicitEnvReferenceName(value string) (string, bool) {
-	if strings.HasPrefix(value, "${") && strings.HasSuffix(value, "}") {
-		name := strings.TrimSuffix(strings.TrimPrefix(value, "${"), "}")
-		return name, isExplicitEnvName(name)
-	}
-	if strings.HasPrefix(value, "$") {
-		name := strings.TrimPrefix(value, "$")
-		return name, isExplicitEnvName(name)
-	}
-	return "", false
-}
-
-func isExplicitEnvName(name string) bool { //nolint:gocognit // baseline (#521)
-	if name == "" {
-		return false
-	}
-	isLetter := func(r rune) bool {
-		return (r >= 'A' && r <= 'Z') || (r >= 'a' && r <= 'z')
-	}
-	for i, r := range name {
-		switch {
-		case i == 0 && (r == '_' || isLetter(r)):
-			continue
-		case i > 0 && (r == '_' || isLetter(r) || (r >= '0' && r <= '9')):
-			continue
-		default:
-			return false
-		}
-	}
-	return true
-}
-
-// NormalizeStateConcurrencyLimits canonicalizes the per-state concurrency
-// cap map (`agent.max_concurrent_agents_by_state`). It is the single
-// source of truth shared between the loader's initial-load path
-// (internal/workflow/loader.go) and the orchestrator's snapshot-build
-// path (internal/orchestrator/workflow_runtime.go); both call this
-// helper so a `WORKFLOW.md` reload cannot produce a differently-shaped
-// in-memory map than the initial load did.
-//
-// Semantics (closes #294):
-//   - Whitespace/case-folded keys: state names are normalized via
-//     [NormalizeStateConcurrencyKey] so `"In Progress"`, `"in progress"`,
-//     and `"  in progress "` all map to the same bucket.
-//   - Empty / whitespace-only keys are DROPPED. The orchestrator looks
-//     up caps by `NormalizeStateConcurrencyKey(stateName)`, which can
-//     never produce the empty string from a real tracker state, so any
-//     preserved empty-key entry would be permanently dead.
-//   - Non-positive limits (`<= 0`) are DROPPED. SPEC §5.3.5 caps are a
-//     positive integer; `0` would silently mean "never dispatch this
-//     state" but operators expressing that intent would set the issue
-//     to a different active-states list, not a `0` cap.
-//
-// Both rules drop entries the orchestrator could not use anyway,
-// trading a little operator-feedback granularity for shape parity
-// across the load/reload boundary.
-func NormalizeStateConcurrencyLimits(limits map[string]int) map[string]int {
-	if len(limits) == 0 {
-		return nil
-	}
-	out := make(map[string]int, len(limits))
-	for state, limit := range limits {
-		key := NormalizeStateConcurrencyKey(state)
-		if key == "" || limit <= 0 {
-			continue
-		}
-		out[key] = limit
-	}
-	return out
-}
-
-// isLinearGraphQLMutationName reports whether s is a valid GraphQL Name
-// per the spec (https://spec.graphql.org/October2021/#sec-Names): an ASCII
-// letter or underscore followed by ASCII letters, digits, or underscores.
-// Used by the codex.linear_graphql.allowed_mutations validator so a typo
-// like " issueUpdate" (with leading space) fails at load time rather than
-// at the first attempted mutation.
-func isLinearGraphQLMutationName(s string) bool { //nolint:gocognit // baseline (#521)
-	if s == "" {
-		return false
-	}
-	for i := 0; i < len(s); i++ {
-		ch := s[i]
-		letter := (ch >= 'A' && ch <= 'Z') || (ch >= 'a' && ch <= 'z')
-		digit := ch >= '0' && ch <= '9'
-		if i == 0 {
-			if !letter && ch != '_' {
-				return false
-			}
-			continue
-		}
-		if !letter && !digit && ch != '_' {
-			return false
-		}
-	}
-	return true
-}
-
-// NormalizeStateConcurrencyKey canonicalizes a single tracker state name
-// for use as a key in the per-state concurrency cap map. The shape
-// (`strings.ToLower` + trim + space→underscore) matches how the
-// orchestrator looks up caps when a worker session is dispatching, so
-// the keys produced by [NormalizeStateConcurrencyLimits] line up with
-// the runtime lookup path.
-func NormalizeStateConcurrencyKey(state string) string {
-	return strings.ReplaceAll(strings.ToLower(strings.TrimSpace(state)), " ", "_")
 }
